fix and add for task 9!

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -11,17 +11,17 @@
  */
 int interpolation_search(int *array, size_t size, int value)
 {
-	unsigned int pos, low = 0, high = size - 1;
+	size_t pos, low = 0, high = size - 1;
 
 	if (array == NULL)
 		return (-1);
 
 	while (size > 0)
 	{
-		pos = low + (((double)(high - low) /
+		pos = (size_t)low + (((double)(high - low) /
 					  (array[high] - array[low])) * (value - array[low]));
 
-		printf("Value checked array[%d]", pos);
+		printf("Value checked array[%d]", (int)pos);
 		if (pos > size)
 		{
 			printf(" is out of range\n");
